refactor: Tab hook for logic (#76)
